Revert "Navbar"
