--- client/src/Marks/Marks.js
@@ -1,12 +1,0 @@
-import React from 'react'
-
-export default function () {
-  return (
-    <div>
-    <h1> Marks will be found here</h1>
-        <p>
-        Sed facilisis tempor elit eget tristique. Sed tincidunt tempor tincidunt. Vestibulum maximus nibh tortor, quis maximus mi consectetur vitae. Pellentesque scelerisque tortor erat, id elementum lacus pulvinar eget. In ultrices nisl sed velit mollis, sit amet dictum magna interdum. Donec quis risus at libero commodo egestas vitae sit amet mi. Ut id neque diam. Nullam lacus risus, hendrerit in lacus et, porttitor vehicula diam. In ultricies feugiat erat ac dignissim. Nullam condimentum lectus eu neque elementum finibus. Maecenas aliquam efficitur sem quis hendrerit. Fusce nisl augue, facilisis ut fermentum eu, gravida vel orci. Morbi leo magna, interdum iaculis efficitur quis, tempor a purus. Nunc nunc quam, convallis vel suscipit quis, egestas tempus erat. Ut a accumsan magna. Nunc porttitor justo suscipit, vehicula sapien vitae, scelerisque mauris.
-        </p>
-    </div>
-  )
-}
