--- client/src/index.css
@@ -1,54 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;700&display=swap');
-
-:root {
-  --primary: #0097b2;
-  --secondary: #808080;
-}
-body {
-  margin: 0;
-  padding: 20px;
-  font-family: "Poppins";
-  background: var(--secondary);
-}
-* {
-  color: white;
-  margin: 0;
-}
-p {
-  margin: 20px 0;
-}
-button {
-  border: 0;
-  padding: 8px;
-  border-radius: 4px;
-  color: white;
-  background: var(--primary);
-  cursor: pointer;
-}
-
-/* navbar */
-header nav {
-  display: flex;
-  gap: 16px;
-  justify-content: end;
-  max-width: 1200px;
-  margin: 0 auto;
-}
-header nav h1 {
-  margin-right: auto;
-  border-bottom: 3px solid var(--primary);
-}
-header nav a{
-  text-decoration: none;
-  padding: 6px;
-  border-radius: 4px;
-}
-header nav a.active {
-  background: var(--primary);
-}
-
-/* page content */
-main {
-  max-width: 1200px;
-  margin: 40px auto;
-}
