+++ client/src/index.js
@@ -1,6 +1,5 @@
 import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
-import "/Users/admin/Desktop/Newsletter-Signup/CodeCrafters-YIT/client/src/index.css";
 
 import App from "./App";
